Move LOD calculation to viewport
