+++ packages/core/src/core/viewport.ts
@@ -1,4 +1,5 @@
 import { Camera } from "../objects/cameras/camera";
+import { OrthographicCamera } from "../objects/cameras/orthographic_camera";
 import { Layer } from "./layer";
 import { LayerManager } from "./layer_manager";
 import { CameraControls } from "../objects/cameras/controls";
@@ -94,9 +95,24 @@ export class Viewport {
     );
   }
 
-  public clientToWorld(position: vec2, depth: number = 0): vec3 {
-    const clipPos = this.clientToClip(position, depth);
-    return this.camera.clipToWorld(clipPos);
+  public getWorldViewRect(): Box2 {
+    if (this.camera.type !== "OrthographicCamera") {
+      throw new Error(
+        "ChunkManager currently supports only orthographic cameras. " +
+          "Update the implementation before using a perspective camera."
+      );
+    }
+    const camera = this.camera as OrthographicCamera;
+    return camera.getWorldViewRect();
+  }
+
+  public getLodFactor(): number {
+    const viewBounds2D = this.getWorldViewRect();
+    const virtualWidth = Math.abs(viewBounds2D.max[0] - viewBounds2D.min[0]);
+    const bufferWidth = this.getBox().toRect().width;
+    const virtualUnitsPerScreenPixel = virtualWidth / bufferWidth;
+
+    return Math.log2(1 / virtualUnitsPerScreenPixel);
   }
 
   private getBox(): Box2 {
